+++ galsim/catalog.py
@@ -22,6 +22,7 @@ Routines for controlling catalog input/output with GalSim.
 from future.utils import iteritems, iterkeys, itervalues
 from builtins import zip
 import galsim
+import numpy as np
 
 class Catalog(object):
     """A class storing the data from an input catalog.
@@ -111,12 +112,11 @@ class Catalog(object):
                     self.nobjects = sum(1 for line in f if not line.startswith(comments))
             return
 
-        import numpy
         # Read in the data using the numpy convenience function
         # Note: we leave the data as str, rather than convert to float, so that if
         # we have any str fields, they don't give an error here.  They'll only give an 
         # error if one tries to convert them to float at some point.
-        self.data = numpy.loadtxt(self.file_name, comments=comments, dtype=bytes)
+        self.data = np.loadtxt(self.file_name, comments=comments, dtype=bytes)
         # Convert the bytes to str.  For Py2, this is a no op.
         self.data = self.data.astype(str)
 
@@ -466,19 +466,17 @@ class OutputCatalog(object):
     def makeData(self):
         """Returns a numpy array of the data as it should be written to an output file.
         """
-        import numpy
-
         cols = zip(*self.rows)
 
         dtypes = []
         new_cols = []
         for col, name, t in zip(cols, self.names, self.types):
             name = str(name)  # numpy will barf if the name is a unicode string
-            dt = numpy.dtype(t) # just used to categorize the type into int, float, str
-            if dt.kind in numpy.typecodes['AllInteger']:
+            dt = np.dtype(t) # just used to categorize the type into int, float, str
+            if dt.kind in np.typecodes['AllInteger']:
                 dtypes.append( (name, int) )
                 new_cols.append(col)
-            elif dt.kind in numpy.typecodes['AllFloat']:
+            elif dt.kind in np.typecodes['AllFloat']:
                 dtypes.append( (name, float) )
                 new_cols.append(col)
             elif t == galsim.Angle:
@@ -501,13 +499,13 @@ class OutputCatalog(object):
                 new_cols.append( [ val.g2 for val in col ] )
             else:
                 col = [ str(s).encode() for s in col ]
-                maxlen = numpy.max([ len(s) for s in col ])
+                maxlen = np.max([ len(s) for s in col ])
                 dtypes.append( (name, str, maxlen) )
                 new_cols.append(col)
 
-        data = numpy.array(list(zip(*new_cols)), dtype=dtypes)
+        data = np.array(list(zip(*new_cols)), dtype=dtypes)
 
-        sort_index = numpy.argsort(self.sort_keys)
+        sort_index = np.argsort(self.sort_keys)
         data = data[sort_index]
 
         return data
@@ -518,8 +516,6 @@ class OutputCatalog(object):
         @param file_name    The name of the file to write to.
         @param prec         Output precision for floats. [default: 8]
         """
-        import numpy
-
         data = self.makeData()
 
         width = prec+8
@@ -531,20 +527,20 @@ class OutputCatalog(object):
         fmt = []
         for name in data.dtype.names:
             dt = data.dtype[name]
-            if dt.kind in numpy.typecodes['AllInteger']:
+            if dt.kind in np.typecodes['AllInteger']:
                 fmt.append('%%%dd'%(width))
-            elif dt.kind in numpy.typecodes['AllFloat']:
+            elif dt.kind in np.typecodes['AllFloat']:
                 fmt.append('%%%d.%de'%(width,prec))
             else:
                 fmt.append('%%%ds'%(width))
 
         try:
-            numpy.savetxt(file_name, data, fmt=fmt, header=header)
+            np.savetxt(file_name, data, fmt=fmt, header=header)
         except (AttributeError, TypeError):
             # header was added with version 1.7, so do it by hand if not available.
             with open(file_name, 'w') as fid:
                 fid.write('#' + header + '\n')
-                numpy.savetxt(fid, data, fmt=fmt)
+                np.savetxt(fid, data, fmt=fmt)
 
     def writeFits(self, file_name):
         """Write catalog to a FITS file.
@@ -562,7 +558,6 @@ class OutputCatalog(object):
         # Note to developers: Because of problems with pickling in older pyfits versions, this
         # code is duplicated in galsim/config/extra_truth.py, BuildTruthHDU.  If you change
         # this function, you should update BuildTruthHDU as well.
-        import numpy
         from galsim._pyfits import pyfits
 
         data = self.makeData()
@@ -570,9 +565,9 @@ class OutputCatalog(object):
         cols = []
         for name in data.dtype.names:
             dt = data.dtype[name]
-            if dt.kind in numpy.typecodes['AllInteger']:
+            if dt.kind in np.typecodes['AllInteger']:
                 cols.append(pyfits.Column(name=name, format='J', array=data[name]))
-            elif dt.kind in numpy.typecodes['AllFloat']:
+            elif dt.kind in np.typecodes['AllFloat']:
                 cols.append(pyfits.Column(name=name, format='D', array=data[name]))
             else:
                 cols.append(pyfits.Column(name=name, format='%dA'%dt.itemsize, array=data[name]))