+++ galsim/des/des_meds.py
@@ -25,9 +25,10 @@ Importing this module also adds these data structures to the config framework, s
 output can subsequently be simulated directly using a config file.
 """
 
-import numpy
+import numpy as np
 import galsim
 import galsim.config
+import sys
 
 # these image stamp sizes are available in MEDS format
 BOX_SIZES = [32,48,64,96,128,192,256]
@@ -176,8 +177,6 @@ def WriteMEDS(obj_list, file_name, clobber=True):
                          existing files. (Default `clobber = True`.)
     """
 
-    import numpy
-    import sys
     from galsim._pyfits import pyfits
 
     # initialise the catalog
@@ -215,14 +214,14 @@ def WriteMEDS(obj_list, file_name, clobber=True):
     for obj in obj_list:
 
         # initialise the start indices for each image
-        start_rows = numpy.ones(MAX_NCUTOUTS)*EMPTY_START_INDEX
-        psf_start_rows = numpy.ones(MAX_NCUTOUTS)*EMPTY_START_INDEX
-        dudrow = numpy.ones(MAX_NCUTOUTS)*EMPTY_JAC_diag
-        dudcol = numpy.ones(MAX_NCUTOUTS)*EMPTY_JAC_offdiag
-        dvdrow = numpy.ones(MAX_NCUTOUTS)*EMPTY_JAC_offdiag
-        dvdcol = numpy.ones(MAX_NCUTOUTS)*EMPTY_JAC_diag
-        row0   = numpy.ones(MAX_NCUTOUTS)*EMPTY_SHIFT
-        col0   = numpy.ones(MAX_NCUTOUTS)*EMPTY_SHIFT
+        start_rows = np.ones(MAX_NCUTOUTS)*EMPTY_START_INDEX
+        psf_start_rows = np.ones(MAX_NCUTOUTS)*EMPTY_START_INDEX
+        dudrow = np.ones(MAX_NCUTOUTS)*EMPTY_JAC_diag
+        dudcol = np.ones(MAX_NCUTOUTS)*EMPTY_JAC_offdiag
+        dvdrow = np.ones(MAX_NCUTOUTS)*EMPTY_JAC_offdiag
+        dvdcol = np.ones(MAX_NCUTOUTS)*EMPTY_JAC_diag
+        row0   = np.ones(MAX_NCUTOUTS)*EMPTY_SHIFT
+        col0   = np.ones(MAX_NCUTOUTS)*EMPTY_SHIFT
 
         # get the number of cutouts (exposures)
         n_cutout = obj.n_cutouts
@@ -282,10 +281,10 @@ def WriteMEDS(obj_list, file_name, clobber=True):
         cat['col0'].append(col0)
 
     # concatenate list to one big vector
-    vec['image'] = numpy.concatenate(vec['image'])
-    vec['seg'] = numpy.concatenate(vec['seg'])
-    vec['weight'] = numpy.concatenate(vec['weight'])
-    vec['psf'] = numpy.concatenate(vec['psf'])
+    vec['image'] = np.concatenate(vec['image'])
+    vec['seg'] = np.concatenate(vec['seg'])
+    vec['weight'] = np.concatenate(vec['weight'])
+    vec['psf'] = np.concatenate(vec['psf'])
 
     # get the primary HDU
     primary = pyfits.PrimaryHDU()
@@ -302,7 +301,7 @@ def WriteMEDS(obj_list, file_name, clobber=True):
     cols.append( pyfits.Column(name='file_id',        format='%dK' % MAX_NCUTOUTS,
                                array=[1]*n_obj) )
     cols.append( pyfits.Column(name='start_row',      format='%dK' % MAX_NCUTOUTS,
-                               array=numpy.array(cat['start_row'])) )
+                               array=np.array(cat['start_row'])) )
     cols.append( pyfits.Column(name='orig_row',       format='%dD' % MAX_NCUTOUTS,
                                array=[[0]*MAX_NCUTOUTS]*n_obj     ) )
     cols.append( pyfits.Column(name='orig_col',       format='%dD' % MAX_NCUTOUTS,
@@ -312,20 +311,20 @@ def WriteMEDS(obj_list, file_name, clobber=True):
     cols.append( pyfits.Column(name='orig_start_col', format='%dK' % MAX_NCUTOUTS,
                                array=[[0]*MAX_NCUTOUTS]*n_obj     ) )
     cols.append( pyfits.Column(name='cutout_row',     format='%dD' % MAX_NCUTOUTS,
-                               array=numpy.array(cat['row0'])     ) )
+                               array=np.array(cat['row0'])     ) )
     cols.append( pyfits.Column(name='cutout_col',     format='%dD' % MAX_NCUTOUTS,
-                               array=numpy.array(cat['col0'])     ) )
+                               array=np.array(cat['col0'])     ) )
     cols.append( pyfits.Column(name='dudrow',         format='%dD' % MAX_NCUTOUTS,
-                               array=numpy.array(cat['dudrow'])   ) )
+                               array=np.array(cat['dudrow'])   ) )
     cols.append( pyfits.Column(name='dudcol',         format='%dD' % MAX_NCUTOUTS,
-                               array=numpy.array(cat['dudcol'])   ) )
+                               array=np.array(cat['dudcol'])   ) )
     cols.append( pyfits.Column(name='dvdrow',         format='%dD' % MAX_NCUTOUTS,
-                               array=numpy.array(cat['dvdrow'])   ) )
+                               array=np.array(cat['dvdrow'])   ) )
     cols.append( pyfits.Column(name='dvdcol',         format='%dD' % MAX_NCUTOUTS,
-                               array=numpy.array(cat['dvdcol'])   ) )
+                               array=np.array(cat['dvdcol'])   ) )
     cols.append( pyfits.Column(name='psf_box_size',   format='K', array=cat['psf_box_size'] ) )
     cols.append( pyfits.Column(name='psf_start_row',  format='%dK' % MAX_NCUTOUTS,
-                               array=numpy.array(cat['psf_start_row'])) )
+                               array=np.array(cat['psf_start_row'])) )
 
 
     # Depending on the version of pyfits, one of these should work: